Move `onKeyDown` logic to swatch and delete new component.
